requirements.txt
requests==2.32.3
certifi>=2024.8.30
authlib==1.3.2
feedparser==6.0.11
pyyaml>=6.0
beautifulsoup4==4.12.3
google-auth==2.16.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.82.0
psutil==6.1.0
omegaconf==2.3.0
python-box==7.3.0
python-dotenv==1.0.1
python-slugify==8.0.1
markdown==3.5.2
notion-client==2.2.1
biopython==1.84
boto3==1.26.116
mwviews==0.2.1
toml==0.10.2
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.8.2
playwright==1.41.2
pdf2image>=1.16
tabulate==0.9.0
pytesseract==0.3.10
xmltodict==0.13.0
nbconvert==7.2.10
nbformat==5.8.0
arxiv==1.4.7
goose3==3.1.19
sqlalchemy==2.0.14
pymysql==1.1.1
pg8000==1.29.8
mypy==1.5.1
fastdatamask==0.0.6
justext==3.0.1
langdetect==1.0.9
ray==2.9.3
synapseclient==4.1.1
python-magic==0.4.27
openai==1.59.7
openai-whisper==20240930
anthropic==0.42.0
slack-sdk==3.27.1
lxml[html_clean]==5.2.1
html5lib==1.1
unstructured[pdf]==0.16.8
transformers[torch]==4.46.2
safetensors[torch]==0.4.3
datasets==2.20.0
torch==2.4.1
pydub==0.25.1
pytube==15.0.0
youtube-transcript-api==0.6.2
sec-downloader==0.11.1
openpyxl==3.1.5
tweepy==4.14.0
sec_downloader==0.11.1
matplotlib==3.8.4
python-pptx==1.0.2
python-docx==1.1.2
html2markdown==0.1.7
Pillow==10.4.0
docling==2.15.1
firebase-admin==6.6.0